all_numeric {visdat}	R Documentation

(Internal) Are they all numeric columns?

Description

(Internal) Are they all numeric columns?

Usage

all_numeric(x, ...)

Arguments

x
data.frame

...
optional extra inputs

Value

logical - TRUE means that there is a column with numerics, FALSE means that there is a column that is not numeric

Examples

## Not run: 
all_numeric(airquality) # TRUE
all_numeric(iris) # FALSE

## End(Not run)

[Package visdat version 0.5.3 Index]