compare_print {visdat}	R Documentation

(Internal) A utility function for vis_compare

Description

compare_print is an internal function that takes creates a dataframe with information about where there are differences in the dataframe. This function is used in vis_compare. It evaluates on the data (df1 == df2) and (currently) replaces the "true" (the same) with "Same" and FALSE with "Different", unless it is missing (coded as NA), in which case it leaves it as NA.

Usage

compare_print(x)

Arguments

x
a vector

[Package visdat version 0.5.3 Index]