miss_guide_label {visdat}	R Documentation

Label the legend with the percent of missing data

Description

miss_guide_label is an internal function for vis_miss to label the legend.

Usage

miss_guide_label(x)

Arguments

x
is a dataframe passed from vis_miss(x).

Value

a tibble with two columns p_miss_lab and p_pres_lab, containing the labels to use for present and missing. A dataframe is returned because I think it is a good style habit compared to a list.

[Package visdat version 0.5.3 Index]