A small toy dataset of imaginary people

A wider dataset than typical_data containing information about some randomly generated people, created using the excellent wakefield package. It is created as deliberately odd / eclectic dataset.

typical_larger_data

Format

A data frame with 300 rows and 49 variables:

Age

Age of each individual, see ?wakefield::age for more info

Animal

A vector of animals, see ?wakefield::animal

Answer

A vector of "Yes" or "No"

Area

A vector of living areas "Suburban", "Urban", "Rural"

Car

names of cars - see ?mtcars

Children

vector of number of children - see ?wakefield::children

Coin

character vector of "heads" and "tails"

Color

vector of vectors from "colors()"

Date

vector of "important" dates for an individual

Death

TRUE / FALSE for whether this person died

Dice

6 sided dice result

DNA

vector of GATC nucleobases

DOB

birth dates

Dummy

a 0/1 dummy var

Education

education attainment level

Employment

employee status

Eye

eye colour

Grade

percent grades

Grade_Level

favorite school grade

Group

control or treatment

hair

hair colours - "brown", "black", "blonde", or "red"

Height

height in cm

Income

yearly income

Browser

choice of internet browser

IQ

intelligence quotient

Language

random language of the world

Level

levels between 1 and 4

Likert

likert response - "strongly agree", "agree", and so on

Lorem_Ipsum

lorem ipsum text

Marital

marital status- "married", "divorced", "widowed", "separated", etc

Military

miliary branch they are in

Month

their favorite month

Name

their name

Normal

a random normal number

Political

their favorite political party

Race

their race

Religion

their religion

SAT

their SAT score

Sentence

an uttered sentence

Sex_1

sex of their first child

Sex_2

sex of their second child

Smokes

do they smoke

Speed

their median speed travelled in a car

State

the last state they visited in the USA

String

a random string they smashed out on the keyboard

Upper

the last key they hit in upper case

Valid

TRUE FALSE answer to a question

Year

significant year to that individuals

Zip

a zip code they have visited

Note

the following code was used to create this data

library(wakefield)

set.seed(1214) typical_larger_data <- r_data_theme(n = 300, data_theme = "the_works")