vis_cor {visdat}	R Documentation

Visualise correlations amongst variables in your data as a heatmap

Description

Visualise correlations amongst variables in your data as a heatmap

Usage

vis_cor(data, cor_method = "pearson",
  na_action = "pairwise.complete.obs", ...)

Arguments

data
data.frame

cor_method
correlation method to use, from cor: "a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated."

na_action
The method for computing covariances when there are missing values present. This can be "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs" (default). This option is taken from the cor function argument use.

...
extra arguments you may want to pass to cor

Value

ggplot2 object

Examples

vis_cor(airquality)
## Not run: 
vis_cor(mtcars)
vis_cor(iris)

## End(Not run)

[Package visdat version 0.5.3 Index]