vis_extract_value_ {visdat}	R Documentation

(Internal) Add values of each row as a column

Description

This adds information about each row, so that when called by plotly, the values are made visible on hover. Warnings are suppressed because tidyr gives a warning about type coercion, which is fine.

Usage

vis_extract_value_(x)

Arguments

x
dataframe created from vis_gather_

Value

the x dataframe with the added column value.

[Package visdat version 0.5.3 Index]