vis_gather_ {visdat}	R Documentation

(Internal) Gather rows into a format appropriate for grid visualisation

Description

(Internal) Gather rows into a format appropriate for grid visualisation

Usage

vis_gather_(x)

Arguments

x
a dataframe

Value

data.frame gathered to have columns "variables", "valueType", and a row id called "rows".

[Package visdat version 0.5.3 Index]