visdat

visdat is a package that helps with the preliminary visualisation of data. visdat makes it easy to visualise your whole dataset so that you can visually identify problems.

See also

Author

Maintainer: Nicholas Tierney nicholas.tierney@gmail.com (ORCID)

Other contributors:

Sean Hughes (ORCID) (Sean Hughes reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/87) [reviewer]

Mara Averick (Mara Averick reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/87) [reviewer]

Stuart Lee [contributor]

Earo Wang [contributor]

Nic Crane [contributor]

Christophe Regouby [contributor]