add_vis_dat_pal {visdat}	R Documentation

(Internal) Add a specific palette to a visdat plot

Description

(Internal) Add a specific palette to a visdat plot

Usage

add_vis_dat_pal(vis_plot, palette)

Arguments

vis_plot
visdat plot created using vis_gather_, vis_extract_value and vis_create_

palette
character "default", "qual" or "cb_safe". "default" (the default) provides the stock ggplot scale for separating the colours. "qual" uses an experimental qualitative colour scheme for providing distinct colours for each Type. "cb_safe" is a set of colours that are appropriate for those with colourblindness. "qual" and "cb_safe" are drawn from http://colorbrewer2.org/.

Value

a visdat plot with a particular palette

Examples

## Not run: 
# see internal use inside vis_guess and vis_dat

## End(Not run)

[Package visdat version 0.5.3 Index]