guess_type {visdat}	R Documentation

(Internal) Guess the type of each individual cell in a dataframe

Description

vis_guess uses guess_type to guess cell elements, like fingerprint.

Usage

guess_type(x)

Arguments

x
is a vector of values you want to guess

Value

a character vector that describes the suspected class. e.g., "10" is an integer, "20.11" is a double, "text" is character, etc.

Examples

## Not run: 
guess_type(1)

guess_type("x")

guess_type(c("1", "0L"))

purrr::map_df(iris, guess_type)

## End(Not run)

[Package visdat version 0.5.3 Index]