A small toy dataset of imaginary people

A dataset containing information about some randomly generated people, created using the excellent wakefield package. It is created as deliberately messy dataset.

typical_data

Format

A data frame with 5000 rows and 11 variables:

ID

Unique identifier for each individual, a sequential character vector of zero-padded identification numbers (IDs). see ?wakefield::id for more info

Race

Race for each individual, "Black", "White", "Hispanic", "Asian", "Other", "Bi-Racial", "Native", and "Hawaiin", see ?wakefield::race for more info

Age

Age of each individual, see ?wakefield::age for more info

Sex

Male or female, see ?wakefield::sex for more info

Height(cm)

Height in centimeters, see ?wakefield::height for more info

IQ

vector of intelligence quotients (IQ), see ?wakefield::iq for more info

Smokes

whether or not this person smokes, see ?wakefield::smokes for more info

Income

Yearly income in dollars, see ?wakefield::income for more info

Died

Whether or not this person has died yet., see ?wakefield::died for more info

Note

the following code was used to create this data library(wakefield)

set.seed(1145)

typical_data <- r_data_frame(n = 5000, id, race, age, sex, height_cm, iq, smokes, income, died)