visdat {visdat}	R Documentation

visdat

Description

visdat is a package that helps with the preliminary visualisation of data. visdat makes it easy to visualise your whole dataset so that you can visually identify problems.

See Also

It's main functions are:

vis_dat()

vis_miss()

vis_guess()

vis_compare()

vis_expect()

Learn more about visdat at www.njtierney.com/visdat/articles/using_visdat.html

[Package visdat version 0.5.3 Index]